Move employee-edit functions into User component
Use two use-cases in to add or edit employee
